Remove `refreshInterval`
I think this is the cause of an intermittent issue where sometimes when adding or removing, the dashboard would revalidate before it finished propagating, causing it to disappear/reappear. I can't find anything for disabling revalidations on interval temporarily, so for now I'm just removing the whole thing. It's nice to have if it could work, but this app isn't intended for live collaboration so it's okay if it only revalidates on focus/reconnect/etc